gododo

+++ ft_strjoin.c
@@ -6,19 +6,18 @@
 /*   By: emaillet <emaillet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/14 14:52:01 by emaillet          #+#    #+#             */
-/*   Updated: 2024/10/16 21:16:02 by emaillet         ###   ########.fr       */
+/*   Updated: 2024/10/16 21:19:02 by emaillet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
-#include <stdlib.h>
 
 char	*ft_strjoin(char const *s1, char const *s2)
 {
 	int		i;
 	char	*dest;
 
-	dest = malloc((ft_strlen(s1) + ft_strlen(s2) + 1) * sizeof(char));
+	dest = malloc(((int)ft_strlen(s1) + (int)ft_strlen(s2) + 1) * sizeof(char));
 	i = 0;
 	while (s1[i])
 	{
